Finished creating dataGenerator for db, renamed table for clarity
